fix timeout update
