+++ mlflow/tracing/trace_manager.py
@@ -4,15 +4,10 @@ import threading
 from dataclasses import dataclass, field
 from typing import Generator, Optional
 
-from cachetools import TTLCache
-
 from mlflow.entities import LiveSpan, Trace, TraceData, TraceInfo
-from mlflow.environment_variables import (
-    MLFLOW_TRACE_BUFFER_MAX_SIZE,
-    MLFLOW_TRACE_BUFFER_TTL_SECONDS,
-    MLFLOW_TRACE_TIMEOUT_SECONDS,
-)
+from mlflow.environment_variables import MLFLOW_TRACE_TIMEOUT_SECONDS
 from mlflow.tracing.constant import SpanAttributeKey
+from mlflow.tracing.utils.timeout import get_trace_cache_with_timeout
 
 _logger = logging.getLogger(__name__)
 
@@ -59,19 +54,9 @@ class InMemoryTraceManager:
         return cls._instance
 
     def __init__(self):
-        from mlflow.tracing.utils.cache import MLflowTraceTimeoutCache
-
         # In-memory cache to store request_id -> _Trace mapping.
-        if MLFLOW_TRACE_TIMEOUT_SECONDS.get() is not None:
-            self._traces = MLflowTraceTimeoutCache(
-                timeout=MLFLOW_TRACE_TIMEOUT_SECONDS.get(),
-                maxsize=MLFLOW_TRACE_BUFFER_MAX_SIZE.get(),
-            )
-        else:
-            self._traces = TTLCache(
-                ttl=MLFLOW_TRACE_BUFFER_TTL_SECONDS.get(),
-                maxsize=MLFLOW_TRACE_BUFFER_MAX_SIZE.get(),
-            )
+        self._traces = get_trace_cache_with_timeout()
+
         # Store mapping between OpenTelemetry trace ID and MLflow request ID
         self._trace_id_to_request_id: dict[int, str] = {}
         self._lock = threading.Lock()  # Lock for _traces
@@ -84,6 +69,8 @@ class InMemoryTraceManager:
             trace_id: The trace ID for the new trace.
             trace_info: The trace info object to be stored.
         """
+        # Check for a new timeout setting whenever a new trace is created.
+        self._check_timeout_update()
         with self._lock:
             self._traces[trace_info.request_id] = _Trace(trace_info)
             self._trace_id_to_request_id[trace_id] = trace_info.request_id
@@ -171,6 +158,25 @@ class InMemoryTraceManager:
             request_id = self._trace_id_to_request_id.pop(trace_id, None)
             trace = self._traces.pop(request_id, None)
         return trace.to_mlflow_trace() if trace else None
+
+    def _check_timeout_update(self):
+        """
+        TTL/Timeout may be updated by users after initial cache creation. This method checks
+        for the update and create a new cache instance with the updated timeout.
+        """
+        new_timeout = MLFLOW_TRACE_TIMEOUT_SECONDS.get()
+        if new_timeout != getattr(self._traces, "timeout", None):
+            if len(self._traces) > 0:
+                _logger.warning(
+                    f"The timeout of the trace buffer has been updated to {new_timeout} seconds. "
+                    "This operation discards all in-progress traces at the moment. Please make "
+                    "sure to update the timeout when there are no in-progress traces."
+                )
+
+            with self._lock:
+                # We need to check here again in case this method runs in parallel
+                if new_timeout != getattr(self._traces, "timeout", None):
+                    self._traces = get_trace_cache_with_timeout()
 
     @classmethod
     def reset(self):
